requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
requests>=2.32.0
pydantic>=2.8.0
django-environ>=0.11.0